services/user-service/src/requirements.txt
# User Service Requirements - Fixed Version
fastapi==0.104.1
uvicorn[standard]==0.35.0  # Use newer version with [standard] extras
python-jose[cryptography]==3.5.0  # Use newer version, fixed typo
passlib[bcrypt]==1.7.4
python-multipart==0.0.6  # Added version for stability

# Core dependencies
pydantic==2.5.0
boto3==1.29.7
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI endpoints

# Optional: For better development experience
python-dateutil==2.8.2  # For date handling
typing-extensions==4.8.0  # For enhanced type hints